requirements.txt
torch
numpy
pandas
matplotlib
seaborn
networkx
rich
pillow
psycopg
torchvision
lightning[extra]
timm
hydra-core
omegaconf
distinctipy
